requirements.txt
torch
transformers
scikit-learn
pandas
matplotlib
tqdm
pyyaml